requirements.txt
gradio==5.23.2
transformers==4.38.2
torch==2.2.1
requests==2.31.0
numpy
pillow